Add missing X-Timestamp to account GET response
Swift account GET responses have an X-Timestamp header representing
the last modification time of that particular account. pfs_middleware
now has that.

Curiously, Swift does not also include the same information in the
(standard) Last-Modified header the way it does for containers. It's
weird, but we're not here to judge, just to match behavior.