pfs_middleware: add container, object counts for account GET
Swift account GET responses have some headers indicating container and
object counts, both total and on a per-storage-policy basis. This
commit adds those headers to pfs_middleware's account GET
responses. We pretend that all containers are in the default storage
policy and that there are no objects, but at least there are some
values in there.