pfs_middleware: fix conditional object HEAD responses
If-Match, If-None-Match, et cetera are supposed to work on both object
GET and HEAD responses. We were only checking them on GET, not HEAD;
now we do both.

This is really just passing "conditional_response=True" to
swob.Response(). There's nothing else that pfs_middleware has to do.

Added tests for If-Match on object GET since we didn't have any.