Respect container ACLs in COALSCE requests
You need to be able to write the target, and you need to be able to
both read *and* write the components.

While we're at it, require that COALESCE elements not include '.' or
'..' as path components, or have '//' anywhere in the path. You can
still include a leading slash like '/container/subdir/object' if you
like but it is (or seems to be) equivalent to 'container/subdir/object',
with both variants being relative to the root of the account/volume.