Change the refcntpool API for buffers uint64 --> int
The Go Programming language specification makes it clear that len() and
cap() return int values and "The implementation guarantees that the result
always fits into an int.", so no single slice can be larger than an int.

    https://golang.org/ref/spec#Length_and_capacity

It also states that the size and capacity arguments to make() are: "The
size arguments n and m must be of integer type or untyped.  A constant
size argument must be non-negative and representable by a value of
type int."

    https://golang.org/ref/spec#Making_slices_maps_and_channels

Therfore the refcntpool will also use int for the size of RefCntBuf and
related APIs.

Update unit tests to match.