Change Makefile to exit on first failure.
Change most makefile targets, including test, vet, fmt, and install
to fail when the first sub-make target fails.  Previously failures
were ignored.

Change the c-* sub-make targets (c-build, c-clean, c-install, etc.) to
print the directory as they enter and leave it.  That's exactly the
opposite of the way its currently done (descending into directories is
silent) so it may be controversial, but it makes the output more readable
when errors occur.