Start and stop Swift processes cleanly
During CHO testing, we noticed that Swift processes were not always
started and stopped as expected when using start_and_mount_pfs or
unmount_and_stop_pfs.  Even though we stopped Swift there were Swift
processes which were still running.  Additionally, the Swift reaper
process was not enabled and therefore the checkpoint container
was not being cleaned up.

The first part of the fix is to do "swift-init all start" and
"swift-init all stop" for starting and stopping Swift.  This also
starts the Swift reaper.

The second part of the fix is in the functional-tests repository.  It is
to have the kill_proxyfs.sh script do a pkill and kill remaining Swift
processes which exist.