Fix hang when mounting with FUSE
We had a window which would cause a hang.  The window happened when
proxyfsd was SIGKILLed after we initiated the FUSE mount but before
our goroutine to service the FUSE request was up and running.

We ended up with an unkillable proxyfsd process.  This would result
in a "proxyfsd <defunct>" process that is only cleared by rebooting
the system.

The fix is to have a WaitGroup and wait until the our Root() method
is called.  This guarantees that we have received at least one
request from the FUSE driver.