Start and stop Swift processes cleanly and FUSE hang (#65)
* Start and stop Swift processes cleanly

During CHO testing, we noticed that Swift processes were not always
started and stopped as expected when using start_and_mount_pfs or
unmount_and_stop_pfs.  Even though we stopped Swift there were Swift
processes which were still running.  

The first part of the fix is in the functional-tests repository.  It is
to have the kill_proxyfs.sh script do a pkill and kill remaining Swift
processes which exist.

* Fix hang when mounting with FUSE

We had a window which would cause a hang.  The window happened when
proxyfsd was SIGKILLed after we initiated the FUSE mount but before
our goroutine to service the FUSE request was up and running.

We ended up with an unkillable proxyfsd process.  This would result
in a "proxyfsd <defunct>" process that is only cleared by rebooting
the system.

The fix is to have a WaitGroup and wait until the our Root() method
is called.  This guarantees that we have received at least one
request from the FUSE driver.

* Remove dead code