Fix DeleteInodeRec()'s recordTransaction() call
This would have resulted during transaction playback of
an attempt to delete Inode#2 from the HeadHunter table.
Fortunately, this would have failed because there is
extreme little chance of Nonce#2 being used for an inode
(indeed, it is now typically used for the first checkpoint
object#)...

The reason for the magic number "2" is that the const
value for transactionDeleteInodeRec is, in fact, 2...
and it was inadvertantly passed as the inodeNumber in
the call to recordTransaction() by DeleteInodeRec().