Rework event loger (evtlog) start and reconfig to fix issues.
Swap the order that logger and evtlog start so logger starts first.
Otherwise there is not way to log problems with evtlog startup.

Rework evtlog.ExpandAndResume() so it starts event logging even if it
was not previously started.

Create a single function to parse the config map and use it in evtlog.Up()
and evtlog.ExpandAndResume().

Change logger reconfig so logging is not disabled by the call to
logger.PauseAndContract().  Instead, changes are deferred until the
call to logger.ExpandAndResume().