Add a file system "layout report" to the internal HTTP server.
This also checks the layout report against the seperately maintained
layout maps and logs as errors any differences.

Functional change: delete objects from each of the cached Layout maps
when their size becomes zero, instead of waiting for the object to have
zero size in all three of the maps.