Enabled arming halter triggers in pfs-crash
Prior reports of FSCK corruption appear to now be explained by the case
where, even though we request the "trafficScript" be terminated, it may
not be completely terminated. As such, on the next start of proxyfsd,
the trafficScript is actually (still) running. When FSCK is performed,
trafficScript is, again, (still) running. And since FSCK is actually
*NOT* re-entrant w.r.t. I/O, corruption is reported... and, indeed,
caused.

In particular, it is long running `fio` jobs that seem to expose the
condition. But earlier it was actually guaranteed in the "trigger"
case because we forgot to actually terminate trafficScript.

Meanwhile, addition of pfs-crash/touch.sh demonstrates yet another
problem. On particularly CentOS, a mount will fail if the underlying
directory is non-empty (notably, other OSes simply "hide" what was
under the non-empty directory during the lifetime of the mount).

The trafficScript may very well see the underlying directory at some
point (e.g. after ProxyFS is cleanly shutdown) and, hence, the subsequent
FUSE mount may fail.

Resolving these two issues remains a challenge :-(.