Harden pfs-crash to address FUSE dir & post-dismount trafficScript data
Also includes early work ultimately targetted at ensuring trafficScript
is, in fact, completely halted.

During testing, it appeared that restarting pfs-crash or, indeed,
running `mkproxyfs -F` after an abrupt end to the proxyfsd process,
errors would appear. When run against ramswift, a wireshark analysis
revealed that during the the ContainerGET step, the response comes
back using "chunked" Transfer-Encoding. Suprisingly, the AccountGET
came back as a non-chunked transfer and, hence, "worked".

It should be noted that "real" Swift apparently never uses "chunked"
Transfer-Encoding in reponse to AccountGET nor ContainerGET, so the
lack of support for it in package swiftclient is not overly critical.