Completed GC of B+Tree "Objects" in headhunter
At this point, while there are certainly some TODOs in the code,
it would appear very good timing to merge this PR into development.
Further, our testing should be exercising FSCK regularly to detect
cases where corruption has occurred in the file system.

For the time being, you should expect to occassionally see lines
of the form:

   Removed unreferenced headhunter B+Tree "Object" 0xXXXXXXXXXXXXXXXX

This is due to the outstanding bug in sortedmap where not all "nodes"
of a B+Tree are properly discarded and, hence, remain unreferenced
in the headhunter B+Tree "Object" B+Tree. FSCK will now discover
these and remove them... reporting the removal in the Info: section
of the FSCK report (and, of course, in the ProxyFS Log File).