Fix a bug in sortedmap that loses objects.
Fix a bug in sortedmap discardNode() that fails to call the client
callback tree.BPlusTreeCallbacks.DiscardNode() in some circumstances
potentially causing objects to be leaked.

To hit this bug one of two circumstances must occur, either:

    1) the tree needs to be loaded as sortedmap.OldBPlusTree() and the
    nodes not loaded (tree not walked or modified) or

    2) the number of nodes accessed or modified in all trees sharing the
    same sortedmap.BPlusTreeCache must exceed the eviction threshhold
    for the cache.

For ProxyFS, that means that, either 1) a regular or directory being
deleted must have been created before the last file system was mounted
and the file or directory is not accessed before being deleted, or 2)
the regular file or directory being deleted was created more then 10,010
regular files or 10,010 directories ago.

(the above numbers are based on:
DirEntryCacheEvictHighLimit:        10010
FileExtentMapEvictHighLimit:        10010)