FSCK now uses async delete and new "exhaustive" ContainerGet
SwiftClient now loops through the entire Account or Container assembling
the exhaustive list of Containers/Objects using the "?marker=" Query
String on GETs issued to Swift. As such, RamSwift was also updated to
support this Query String.

The Object Deletes in the Checkpoint Container are now
done in parallel (to the extent the non-chunked connection pool allows).

Fixed pfs-crash/b_tree_load.sh to follow the pattern
of other pfs-crash trafficScript's.