Fixed remaining scalability problems with FSCK
Now, the "grab parallelism" step is performed *BEFORE* launching a
goroutine. Previously, the goroutine was used to grab it. This would
be an ok pattern if one is merely trying to limit the number of
such goroutines executing in parallel because too many would be
inefficient... but here, we are actually worried about the sheer
number of goroutines being too great such that just holding that
number of goroutine contexts in (virtual) memory a potential challenge.

Also:
  Added "End Time" in FSCK/SCRUB page table.
  Updated Layout Report (both JSON & HTML)
  Integrated Trigger pages into JS-enabled single page

Finally, pfs-crash now performs the SCRUB *BEFORE* the FSCK since
the SCRUB may have deleted Inodes that are now not found during the
subsequent FSCK.