Fix a fsck bug, trap and print all signals, and log PID.
Fix a bug in fsck that caused sparse files to be fail validation and
their inodes deleted.

Catch all signals that are catchable and print a message when they
are received.  Log a message as to why proxyfsd is exiting in all cases.

Change the logger to print the current PID on each line.  This will
allow us to distinguish log messages generated by proxyfsd from other
processes like mkproxyfs.  Have mkproxyfs log a startup message that
includes its PID.

(We need to get any other processes that use the logger to also print
a message, with their PID, when they start.)

Improve a couple of log messages when inode validation fails during fsck.
Improve some other log messages.