Ignore SIGPIPE and SIGCONT.
Correctly ignore SIGCHLD, SIGURG, and SIGWINCH (the if statement had no
effect before!).

Add SIGPIPE to signals that are ignored because proxyfsd gets hit
with it when it writes to a closed socket, which can happen if an HTTP
client closes it before reading a response to its request (and it other
circumstances as well).

Ignore SIGCONT juts in case someone is using SIGSTOP/SIGCONT on proxyfsd.
Without this SIGSTOP stops the process and SIGCONT starts it, but
SIGCONT is also caught and caues proxyfsd to exit.