Enabled Async Delete of discarded last checkpoint objects
While the per-SnapShot maintenance of lists of objects to
be garbage collected is not yet available, this change
resumes async deletes of objects made unreferenced by the
just completed checkpoint. This is now a part of the
putCheckpoint() operation.

Ultimately, this logic needs to move to the place where
snapshots are deleted... or, at least, needs to only delete
objects created since the last one.

In the meantime, however, this codebase is now no longer a
regression from that of what's on development branch and
can be merged with no ill effects.