Refactored bPlusTreeWrapperStruct-handling code
Previously, the bPlusTreeWrapperStruct-receiving func's would look at
a "tag" and access wrapper-type-specific fields in the volumeViewStruct.
As the number of such wrappers is growing (was 3 for the 3 HH B+Trees,
now will have another two for each volumeView), this was becoming very
unmanageable. It turns out the wrapper-type was easily removed with
the bPlusTreeWrapperStruct-receiving func's able to process only fields
within the received struct (and not the containing volumeViewStruct).