Add NoAuthIPAddr option to swiftclient
Currently the ip of backend swift proxy is hardcorded as 127.0.0.1.
It's not frexible because bind_ip can be changed in swift proxy-server.conf.

Exactly, connecting to localhost is the fast way on the conversation between
proxyfsd and swift proxy server but sometimes users may want to connect outer
swift resource or may want to bind to a specific ip when the proxy has more than
a tcp ethernet.

This patch enable to configure IP address for proxyfsd to connect.