Flushed out headhunter.volumeViewStruct and how they are managed
This check-in documents the semantics of the headhunter.volumeStruct's
liveView/snapShotViewTree semantics. Comments in the body of the
headhunter.volumeViewStruct document how the createdObjects and
deletedObjects trees are used to track object lifetimes and how
such lifetimes are enforced when the last referencing volumeView
is, itself, deleted. Note that this logic includes handling of the
cases where there are no active SnapShots as well as the cases
where one or more SnapShots exist but the lifetime of objects exists
after the most recent one has been taken.