Completed SnapShot deletion
As part of SnapShot deletion, a checkpoint is performed after the
B+Tree's of the being-deleted volumeView have been emptied. In
the checkpoint, the now empty volumeVew will still be present and,
hence, it's B+Tree's will be flushed. At this time, however, the
checkpoint code still only understands V2 checkpoints and does not
flush the (non-live) volumeViews. The net result is that deleting
a volumeView will crash due to the attempt to discard its B+Trees
that contain dirty (root) nodes.

As such, snapshot_phase_3 will be extended to include completion of
the update of the putCheckpoint() code to support V3 checkpoints.
This, of course, means that getCheckpoint() will also have to be
updated to (also) support loading a V3 checkpoint.

Thus, the subsequent snapshot_phase_4 will only include the addition
of the SnapShot "view" code (i.e. support for the /.snapshot/ directory
tree).