Almost complete with getCheckpoint() changes... 3 TODOs
The final element of Phase 3 will be completion of putCheckpoint().
Note that putCheckpoint() will force an upgrade to V3 of the
checkpoint format... whether or not any snapshots have been created.
This isn't strictly required as the getCheckpoint() code must continue
to load V2 checkpoints (at least for now). But it would require some
amount of extra work to enable volumes with no snapshots declared to
continue writing out V2 checkpoints.

Note: This check-in deprecates the support for changing the format
of the checkpoint container's header. The thinking here is that, going
forward, all checkpoint format changes will be exclusively done to the
checkpoint object trailer... some would call this trailer the "superblock".