Don't call time.Now() for each lock op if lock tracking is disabled.
Calling time.Now() each time we get a lock slightly increases the CPU
consumption of a lock operation, so only call time.Now() when lock tracking
is enabled ('TrackedLock.LockCheckPeriod' > 0).

The code in unlockTrack() and rUnlockTrack() assumed that the lock time
is always valid.  Change it to ignore locks where `lockTime` == 0 (which
can happen if lock tracking is enabled after a lock is acquired).

rLockTrack() was adding the RWMutex to the global rwMutexMap even when
the lock watcher was disabled.  Stop doing that.

rUnlockTrack() had duplicate code to free lock tracking information when
it exists for a lock but lock tracking is disabled (which can happen if
the lock is acquired then tracking is active but tracking is disabled
before the lock is released).  Delete the duplicate code.

If the config file has LockCheckPeriod != 0 and LockHoldTimeLimit == 0
then print a warning and set LockCheckPeriod == 0.

lockWatcher() could spuriously report some locks has been held too long
if LockHoldTimeLimit is set to 0 as a result of a reconfig.  Change the
local variable `longestDuration` to be 10 years if LockHoldTimeLimit is 0
(which can only happen during shutdown processing).

Change the warning messages from `unlockTracker()`, `rUnlockTracker()`,
and the `lockWatcher()` to add an extra newline after the first stack trace.
The unit test log parse regular expressions had to be updated to match.

Add a new unit test, `TestStateTransitions()`, that tries to test all
possible combinations of lock tracking states during lock and unlock
operations.  Change `TestReload()` to utilize some of the mechanism that
it added.