Added /.snapshot/ view support
Also switched dirty state to FALSE for paged-in Inodes.

This set of changes ensures all GETs performed by package inode to
fetch a InodeRec, LogSegmentRec, or BPlusTreeObject from package
headhunter use keys containing the desired SnapShotID. This will
ultimately inform headhunter of which volumeView (live or a prior
snapshot) when fetching/returning the queried-for value.