Fix a bug with retry message for chunked puts.
If the first attempt for a Chunked Put operation failed it would be
retried, bbut the first failure did not count against the retry limit
and we would not see a message that the retry succeeded.  Fix that by
moving the fist call to closeHelper() inside the request function and
special casing the retry() logic so its not called on the first attempt.

More explicitly, previously if there was a failure in SendChunk() we
would see the failure in the log::
> time="2018-08-04T03:24:58.549673Z" level=error msg="swiftclient.chunkedPutContext.SendChunk(\"TestAccount/TestContainer/FooBarCopy\") got writeHTTPPutChunk() error" error="writeHTTPPutChunk() simulated error" function=SendChunk goroutine=6 package=swiftclient

but if the _first_ retry succeeded, we would not see a message by
RequestWithRetry() saying that the request had succeded:
time="2018-08-04T03:24:58.803157Z" level=info msg="retry.RequestWithRetry(): swiftclient.chunkedPutContext.Close(\"TestAccount/TestContainer/FooBarCopy\") succeeded after 2 attempts in 0.253 sec" function=RequestWithRetry goroutine=6 package=swiftclient pid=2204

although we would see the message if it succeeded after 3 or more
attempts.

(logger.PanicfWithError()) because one of them started happening (due
to a bug in my changes) and it wasn't very noticable without a panic.