Fix low NonChunkedPool bug in inode tests
The basic issue was three test cases were launching chunked pool
connections by writing to a file... but not flushing the file.
As a result, at the end of the test - where no proper shutdown
is performed - the chunked PUT connection was left "open". There
were three such test cases... meaning that when running the
entire test suite, subsequent tests would certainly be beyond,
say, a limit of two.

Still, the question remains why didn't the starvationCallback()
take care of the situation. This begs for the missing starvation
test case.