Made checkpoints use a non-pool-supplied Chunked PUT Connection
This change is to avoid a case where all available Chunked PUT
Connections allowed (pool size limited) when a checkpoint needs
one. Normally, a starvationCallback() would free up a connection
from one of the Chunked PUTs underway for a FileInode... but what
if the pool size was smaller than the number of Volumes? In this
case, though unlikely, all of the Volumes might be simultaneously
attempting checkpoints and, hence, the starvationCallback() won't
find any connections to free up.