Rejigger chunked put context flushing.
Separate flushing (Close()) of chunked put contexts so the inode lock
is not required.  chunkedPutConnectionPoolStarvationCallback() can
now flush a chunked put connection without needing any inode locks.

Divorce "open log segments", that is "active" chunked put connections
from inodes by removing the openLogSegment field and replacing
it with a file-system wide "cache" (map) of open log segments,
globals.openLogSegmentCache.  if you want to write to a file
(doSendChunk()), then check globals.openLogSegmentCache for a
connection and, if present, use it.  If not, create a new one.
Change doFileInodeDataFlush() to do the necesary clenaup of chunked put
connections in `fileInode.inFlightLogSegmentMap`, including chunked puts
that failed.

Remove the calls to `openLogSegmentLRUTouch()` in `doReadPlan()` and
then delete openLogSegmentLRUTouch() because its no longer used.  Calling
openLogSegmentLRUTouch() in doReadPlan() is likely a race condition
waiting to core dump since the chunked put contxt in question may not be
in the Log Segment LRU in the first place.

Decrease the number of connections in the chunked put connect pool for
the inode and fs unit tests to "2" and "1" respectively.