Fix bug in acquireChunkedConnection() that caused hangs.
Fix a bug that caused us to hang forever if we hit a race condition in
acquireNonChunkedConnection() or acquireChunkedConnection() by reworking
the code to simplify the synchronization, at the expense of "fairness"
when we are in starvation conditions (we should have flushing at higher
levels that works hard to avoid starvation conditions in the first place).

Part of this change eliminated an endless series of calls to the chunked
put connection pool starvation routine; instead each requestor calls
the starvation callback only once (or again if it is woken up and no
connection slots are available).

This introduces the possibility of a hang if the starvation callback finds
no work to do and the inodes (requests) holding chunked put connections
are never flushed.  The correct solution is to introduce a background
flusher for "open log segments" that insures that, even if idle, they are
eventually flushed to storage (preferably before the Swift server gives
up on them and fails the chunked put operation).  We need the chunked put
connection flusher regardless of whether there is a starvation condition,
so this doesn't really add a requirement for new code.

Change chunkedPutConnectionPoolStarvationCallback() to Close() a
chunked put connection in a separate go routine instead of doing
it itself (and blocking).  This decreases the latency of a call to
chunkedPutConnectionPoolStarvationCallback() to nearly zero, but also
means that the condition is probably not resolved by the time it returns.

Collapse the chunked and nonChunked connection pool code into a single set
of functions that operate on a *connectionPoolStruct.  This avoids bugs
being fixed in the chunked (or nonChunked) pool code and not the other.
This also increases unit test coverage by having all connection related
tests test the same set of code.