Change bucket choice; use the bucket index closest to log(2,value).
This reverts commit 07d0106872dadef8ef9cdbe9ab16ad1d97e53c4f.

Change the bucket tables back to the way they were, with a value going
to the bucket that most closely matches its logarithm (i.e. the bucket
index closest to log(2, value) or log(sqrt(2), value)).

Fix errorPct{Likely,Max} to match logarithmic buckets.

Revise the distribution of random number to be uniform.  This cuts
down the Error Percent Likely by quite a bit so revise those as well.