Rework statistic names to allow "reserved" characters.
Replace code that "scrubs" (or canonicalizes) package names, statistic
group names, or statistic names when they are registered and replace
it with code that scrubs them according to the `StatStringFormat`
Sprint() format used to collect them, since the rules for acceptable
names vary depending on the output format -- a JSON output format can
accomodate almost any character in a name while the "human readable"
format (`StatFormatHumanReadable`) does not allow white space, control
characters or other special characters in a name (they are mapped to
underbar (_).

Refactor the unit tests to move tests of this to TestSprintStats().

Rename StatsFormatHumanReadable --> StatFormatHumanReadable