Finished .conf namespace'ing convertion
Previously, various sections in the .conf were entirely user specified.
This might potentially collide with fixed section names (e.g. Cluster).
So this was changed to prefix such section names with the qualifier for
what it is (e.g. Volume, PhysicalContainerLayout, FlowControl, and Peer)
followed by a colon (':').

This change removes support for the prior unqualified/namespace'd format.