Reworked ramswift daemon startup logic
Previously, ramswift daemon was launched and the "parent" would spin
waiting for a bool passed to each daemon was set to true. The Go
Race Detector would flag this as a race condition... even though it
couldn't actually be since: a) the bool was initialized to false
prior to the daemon's launch, b) only the daemon set it to true, and
c) only the "parent" read the bool awaiting it to be true. Still, this
polling can be cleanly/safely replaced with a sync.WaitGroup.

Note that the proxyfsd.Daemon() launch synchronization works the same
way as did ramswift... but the logic of checking for errors will require
a bit more care in replacing this "racey" pattern.