Introduction of new pfsalived Liveness Detector
This commit includes the framework by which pfsalived is configured
and conducts its polling of selected volumes. It does not, however,
actually perform the liveness detection.

To get an idea of how liveness is reported, checkout the JSON response
from the embedded HTTP server inside pfsalived. While on "Peer0" at
localhost (127.0.0.1) a curl of http://127.0.0.1:15432/status will
return something like this:

{
	"Peer0": {
		"Name": "Peer0",
		"PublicIPAddr": "127.0.0.1",
		"PrivateIPAddr": "127.0.0.1",
		"VolumesToWatch": {
			"CommonVolume": {
				"Name": "CommonVolume",
				"State": "unknown",
				"LastCheckTime": "2018-09-14T11:33:00.867629-07:00"
			}
		}
	}
}