Enhance the chunked put tests by sending larger amounts of data, randomizing the chunk boundaries, and perform concurrent chunked puts where upto maxThread (6) chunked puts are running concurrently, with failures.
Add chaosCloseChunkFailureRate to swiftclient so we also simulate
failures on the call to chunked put Context.Close(), which is
where they typically happen in the "real world".

Fix a bug in chaos testing for chunked put connections where
chunkedPutContextStruct.retry() was not incrementing sendChunkRetryCnt
for each chunk that was resent.

The concurrent puts are a bit problematic since the global nature of
the chaos failure counters and indeterminate scheduling of go routines
means that one goroutine could hit the chaos induced failure repeatedly
while the others avoid it, potentially causing the the retry limit for
the goroutine to expire, which would be flagged as a bug when there's
no actual a bug.

To avoid this, I've tuned the retry limits up and the failue rate down
(bigger values) so this doesn't occur in over 10,000 runs in my test
environment.  But if you change the number of iterations or retry
paramaters you might run into this.

The truly bad errors (data courruption) will panic, whereas the others
will show up as mere errors, which may make it easier to distinguish
the problem.

This can only happen in the testChunkedPut()/testObjectWriteVerify() test.

Fix some strings in error returns.