Liveness Detector is now functional... ping response == alive
The framework is now complete... at least as far as using JSON RPC to
talk to proxyfsd is concerned. At this time, a simple RpcPing is sent
and, if replied-to, liveness is reported. Otherwise, the volume is
considered dead. Of course there is nothing "per-volume" about the
RpcPing sequence... nor does this imply access to the underlying
layers - let alone Swift - is functioning. Those enhancements will
come in the next phase.