Add a counter and chunked put checksum verification to swiftclient.
Add code to checksum the chunked put buffers held by a ChunkedPutContext
for retry (resend) purposes when they are passed to SendChunk() and then
verify the checksum everywhere it seems prudent.

This is enabled by setting:
    SwiftClient.ChecksumChunkedPutChunks=true

in the config file or config stings.  The default is false.

Add the:

	swiftclient.ObjectPutCtxtFetchToCloseUsec

statistic, which represents the amount of time the client is holding the
chunked put connection (measured from the return of the ChunkedPutContext
to call to Close()).  Strictly speaking, this also includes time spent in
SendChunk() sending (or queuing) the data to be sent to the Swift server;
this may or may not be what we want.

Wrapped some lines that were more than 120 characters long.