Convert pakcage fs and headhunters from stats to bucketstats.
First cut of moving all fs statistics to bucketstats and most of
headhunter statistics as well.

Pretty much all API functions get a set of buckets to record the amount
of time spent in the function and those that return an error also get
a bucket that counts the number of times an error is returned.

For example, Read() in package fs gets the bucketized statistics ReadUsec
(microsecond spent in each read request), the total statistic ReadErrors
(number of times an error is returned) and a bonus bucketized statistic,
ReadBytes (number of bytes in each read request).

Eliminated package stats from package fs, but not entirely from
headhunter.