Lower #workers in inode stress test to conserve RAM
In several constrained environments (e.g. runway, SAIO VM), a value
of testStressNumWorkers of 25 causes the amount of trapped RAM in
ramswift (embedded in the inode.test program built for the package)
to exceed available RAM (e.g. 6GB for SAIO VM).

Lowering to 10 will still exercise the exhausted Chunked and NonChunked
connection pools and bound the amount of necessary DRAM to something
around 3GB.