Fix SLO to return correct "last-modified-time".
Change Coalesce() to add ctime (attr change time) to return values.

If the response to a Coalesce() request by middleware contains a ctime
value, it is use for "last-modified-time" in the REST response.  If not,
the mtime value is used.  This was added a few months ago as part of a
transition from mtime --> ctime for last-modifed-time.

CoalesceReply in RpcCoalesce() in jrpcfs/middleware.go includes a ctime
field, but it was not being set by anything.  Presumably its return
value was 0 and the middleware python code would then use the value of 0
for last-modified-time.

Plumb ctime as one of the values returned by Coalesce() through several
layers of the stack so it gets to the Python middleware code as it
should.

Note that we don't seem to have any end-to-end tests (swift-client to
Proxyfs) of SLO uploads (a PUT with "multipart-manifest=put") so we
don't have any tests that would detect this.