Avoid use of colons in scheduled snapshot names
SMB Clients normally prohibit the use of colons in file/directory names.
As each scheduled snapshot bears a name of the time it was taken, such
snapshots will appear in the /.snapshot directory with those names.

Prior to this change, the format chosen for the snapshot names was simply
RFC3339 (equivalent to ISO 8601). This format nicely encodes the timezone
in addition to the date and time. Alas, it specifies the use of colons to
separate hours from minutes and minutes from seconds. This presents quite
a challenge for SMB clients given the colon prohibition.

As a workaround, scheduled snapshot names will take the RFC3339-formatted
time and replace colons with dots (periods). This is a nod to the pre-EU
German (DIN) standard practice and avoids the SMB probibited character
set for file/directory names.