Refactor etcd interactions.
Change queries of the etcd database to return a slice of _comparisons_
and changes updates to items in the etcd datbase to return _operations_.
Actual updates to the etcd database consist of _conditionals_ and the
operations, where the operations are performed if, and only, all of the
conditionals (comparisons) are satisfied.  etcd guarantees that all of
the operations in an update are performed or none of them.

If proxyfs code concatenates the comparisons returned from all get
operations that make up a decision to change the database then it insures
the change is not made unless all of the prerequisites are still true.

Change the ProxyFS scheme for database storage to store all the
information about a Volume Group under a single key with the value
JSON encoded.  Refactor the interfaces to match. (The same change has
not yet been made for per node information, but it will be.)

One of the next steps is to make volumes (file systems) "first class"
citizens since, fundamentally, that is where the rubber meets the road.
Volumes are what get mounted, unmounted, failed, and are what the user
actually observes.  Changes to volume groups and, to a lesser extent,
nodes, only manifest themselves as changes to file systems.

Change the VgState type from an int to a string and discard code code
used to convert between the two.  make VgState a "first class" data type
in the shared database.

Change:
    VgInfo --> AllVgInfo
    NodeInfo --> AllNodeInfo

so we can use VgInfo and NodeInfo for information about an individual
node.  (AllVgInfo no longer exists as a type; its now represented as
a map[string]VgInfo.  but i think we really need to change this to
something that is based on UUID rather than user-assigned names.)

Update the SETUP documentation with a little more information about
parallel-ssh and the start of a "cheat sheet" on how to use etcdctl to
update the database state.