Enable goroutine stack tracebacks via HTTP on port 6060.
This enables goroutien stack tracebacks (and profiling?) via HTTP
on port 6060 in ProxyFS (the daemon) and in the inode unit tests.

It also fixes a potential race condition in the swiftclient unit
tests that use chaos testing (where two threads might race while
updating the objectFetchChunkedPutContextCnt global.

This uses an `atomic.AddUint64()`, but note that it is only used if
Chaos testing is enabled, so there's no overhead except for test
cases.