Proxyfs does not commit suicide when noauth proxy restarts.
When Proxyfs tries to open a new connection to the noauth proxy
server with net.DialTCP() and that attempt fails it was calling
logger.FatalfWithError() which caused it to exit.  This is bad because
the noauth proxy can be down due to a restart caused by a config push.

Change openConnection() handling of a failed open attempt from logging
it as a fatal error (and exiting) and instead log a warning and return
an error to the caller.

Modify its callers, acquireChunkedConnection() and
acquireNonChunkedConnection() to check for an error and return it to
their callers.

And modify their callers to check for an error and handle it as they
would any other error performing the operation, which is to wait a bit
and then retry the operation (upto a maximum number of retries).

There's an interesting potential, but non-existent, bug: we cache
connections and reuse them, which means if the noauth proxy dies and
restarts those cached connections can be bad.  If the retry code were
to perform a retry using a cached connection, that would fail even if
the noauth proxy server had come back up, potentially exhausting all
possible retries.

Happily, recent changes for "starved connections" means that retry
will always re-open the connection it was using (creating a brand new
connection to the noauth proxy), so this does not actually happen.

But the pool of cached bad connections will affect the next new attempt
to do something, causing the first attempt to fail and triggering the
retry timeout.  That will be fixed later.