Log a message to syslog when proxyfsd exits.
Proxyfs logs a message to syslog when it starts up.  Now it logs a message
when it exits as well, with an error indication if an error occurred.

Unfortunately, it will not log anything to syslog _via this mechanism_
if it dies in a way that does not tell the main thread, though go seems
to place SIGSEGV errors into syslog separately.